Game.Types

Documentation

data BoardCoordinate

position of a cell on the board

Constructors

BoardCoordinate
BoardCoordinate column row

Fields

boardColumn :: Int

column (x-axis)

boardRow :: Int

row (y-axis)

data GameAction

a move in the game played by the player

Constructors

Reveal BoardCoordinate
reveal a single cell

RevealArea BoardCoordinate
reveal all adjacent safe cells if possible

ToggleFlag BoardCoordinate
toggle flag/unsure on cell

Undo
undo last Reveal / RevealArea action

data GameState

represents the complete and consistent state of the game in a given point in time

Constructors

GameState
GameState config Maybe previousState cellStates globalGameState

Fields

gameConfig :: GameConfig

board configuration for this game

previousState :: Maybe GameState

previous game state used for undo operations

cells :: CellStates

state of all cells on the board

globalGameState :: GlobalGameState

some global states calculated from the states of all cells

undo :: GameState -> GameState

return the previous game state to undo last reveal

if no previous state is available, the same state is returned

data GameConfig

Board configuration for a new game

Constructors

GameConfig
GameConfig width height numMines

Fields

boardWidth :: Int

number of columns on the board

boardHeight :: Int

number of rows on the board

totalMines :: Int

total number of mines in the game

totalCells :: GameConfig -> Int

total number of cells for this GameConfig

data PlayState

current win/lose state of the game

Constructors

Playing
game is still running

Win
the player won

Dead
the player lost

data GlobalGameState

global states of the game calculated from all cell states

Constructors

GlobalGameState
GlobalGameState remainingMines freeCells playState

Fields

remainingMines :: Int

number of floating mines

freeCells :: Int

number of floating safe cells

globalPlayState :: PlayState

whether game was won/lost or is still playing

data InternalCellState

internal (non-visible) state of a cell

Constructors

Safe
the cell does not contain a mine

Mine
the cell does contain a mine

Undefined
it is not fixed yet whether this cell will contain a mine

data VisibleCellState

state of a cell visible to the player

Constructors

Unknown
cell is still hidden

Known
cell is open (and contains a mine)

Flagged
cell was falgged by the player

Unsure
cell was flagged as unsure by the player

Labeled
cell was revealed and shows the number of adjacent mines

Fields

totalLabel :: Int

total number of adjacent mines

countdownLabel :: Int

remaining number of adjacent mines based on set flags

data CellState

the state of cell on the board

Constructors

CellState
CellState internalState visibleState

Fields

internalCellState :: InternalCellState

internal state of the cell

visibleCellState :: VisibleCellState

visible state of the cell

type CellStates = [[CellState]]

rows and columns of cells representing the game board

isMine :: CellState -> Bool

whether a cell (definitely) contains a mine

isSafe :: CellState -> Bool

whether a cell (definitely) contains no mine

isUndefined :: CellState -> Bool

whether the content of the cell was not fixed yet

isFlagged :: CellState -> Bool

whether the cell was flagged by the player

isKnown :: CellState -> Bool

whether the cell is revealed

cellFromState :: BoardCoordinate -> CellStates -> CellState

get cell on specific coordinate from cellStates

countCells :: (CellState -> Bool) -> [CellState] -> Int

count cells in list matching the predicate

countInState :: (CellState -> Bool) -> CellStates -> Int

count cells in cellStates matching the predicate